Integrate HiSi video decoder support
This adds support for HiSi's video decoder support.

This commit does not contain the library blobs, as those
will be provided in a separate vendor package.

However this seems to gracefully fall back to software
decoding if the libraries aren't present or the kernel
doesn't support the vdec driver (like w/ 4.14+).

Change-Id: I5656e8616b18302a89fb2f708801bd5eb988aa7e
Signed-off-by: John Stultz <john.stultz@linaro.org>