hikey/hikey960: Add more services for treble
In preparation for the transition to FULL_TREBLE,
add some missing services to the product packages.

Credit for this patch goes to Vishal Bhoj <vishal.bhoj@linaro.org>
who basically was the original author and magician
who got things working.

I've only refactored and split up his changes to help
with the submissions.

Change-Id: I8adb8f0c1f76bfdda6efd85446ba6daf7ce58fcc
Signed-off-by: John Stultz <john.stultz@linaro.org>