Enable FULL_TREBLE on hikey/hikey960
This is basically the current state of the rest of the
FULL_TREBLE enablement bits for both hikey/hikey960

This patch also requires kernels that are rebuilt to include the
DTS fstab changes seen here:
https://git.linaro.org/people/john.stultz/android-dev.git/log/?h=dev/treble

Credit for this patch goes to Vishal Bhoj <vishal.bhoj@linaro.org>
who basically was the original author and magician
who got things working.

I've only refactored and split up his changes to help
with the submissions.

Change-Id: Iab0f5fc93bdfd1b86c98aa1d54974fde33514063
Signed-off-by: John Stultz <john.stultz@linaro.org>