hikey32/64: Be sure to include mali blobs in hikey64_only and hikey32 targets
When we moved to the r9p0 binary update, the Android.mk
TARGET_DEVICE filter was too narrow and ended up skipping over
hikey64 and hikey32.

This results in the hikey64_only/hikey32 target not booting to UI, as
its missing the egl library.

This patch fixes the filtering and allows the target to boot to UI.

Bug: 80063321
Change-Id: Ia72aae2324fdba17ba7aa42417271a0d727d8fe4
Signed-off-by: John Stultz <john.stultz@linaro.org>