HACK: Workaround hikey64_only target not building with FULL_TREBLE
Unfortunately, the hikey64_only target doesn't build with
FULL_TREBLE as the soft-codecs only build for 32bit and are
required in the current VNDK-core.

Thus, this patch temporarily disables TREBLE on the
hikey64_only target, so that things continue to build.

Once we have 64bit soft-codecs, then we can remove this
extra conditional.

Change-Id: Id7444c23b6e1faedeb163d65fdb1202be04b7302
Signed-off-by: John Stultz <john.stultz@linaro.org>