hikey960: Disable ARM renderscript driver
It is not compatible with treble as it needs additional dependencies to
be built and installed to /vendor. There is also a conflicting rule to
build android.hardware.renderscript@1.0.so and we are about to
deprecate support for this.

build/make/core/base_rules.mk:294: warning: ignoring old commands for target
`out/target/product/hikey960/obj_arm/lib/android.hardware.renderscript@1.0-impl.so'

Bug: 77611511
Bug: 77909036
Bug: 110479079
Change-Id: I4c0f6643c5f9000c6a2131fd6b1d4170fe8b07f4
Signed-off-by: Alistair Strachan <astrachan@google.com>